refactor: rename theme related files
