+++ src/module.ts
@@ -14,7 +14,7 @@ import { defu } from 'defu'
 import { name, version } from '../package.json'
 import iconifyVitePlugin from './build/iconify'
 import type { PublicConfig } from './runtime/types'
-import { generateThemeConfig } from './runtime/utils'
+import { mergeThemeConfig } from './runtime/themes/merge'
 
 // Module options TypeScript inteface definition
 export interface ModuleOptions extends PublicConfig { }
@@ -36,7 +36,7 @@ export default defineNuxtModule<ModuleOptions>({
     iconSize: 20,
     iconDownload: false,
     iconCollectionsUrl: 'https://iconify-icon-sets.netlify.app',
-    themeConfig: generateThemeConfig(),
+    themeConfig: {},
   },
 
   // Add types for volar
@@ -64,7 +64,7 @@ export default defineNuxtModule<ModuleOptions>({
         iconCollectionsUrl: options.iconCollectionsUrl,
         iconSize: options.iconSize,
         spaLoadingTemplate: options.spaLoadingTemplate,
-        themeConfig: options.themeConfig,
+        themeConfig: mergeThemeConfig(options.themeConfig),
       },
     })
