+++ src/runtime/themes/merge.ts
@@ -1,10 +1,10 @@
 import { defu } from 'defu'
 import type { ThemeConfig } from '../types'
-import defaultLightTheme from '../theme/light'
-import defaultDarkTheme from '../theme/dark'
-import defaultMobileOrTabletTheme from '../theme/mobileOrTablet'
+import defaultLightTheme from './light'
+import defaultDarkTheme from './dark'
+import defaultMobileOrTabletTheme from './mobileOrTablet'
 
-export function generateThemeConfig(themeConfig?: ThemeConfig): ThemeConfig {
+export function mergeThemeConfig(themeConfig?: ThemeConfig): ThemeConfig {
   return {
     dark: themeConfig?.dark?.defaults === false ? themeConfig.dark : defu(themeConfig?.dark, defaultDarkTheme),
     light: themeConfig?.light?.defaults === false ? themeConfig.light : defu(themeConfig?.light, defaultLightTheme),
