+++ src/runtime/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './antd'
 export * from './tailwindcss'
-export * from './theme'
